requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
environ==1.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
